+++ packages/eip6963/src/emitter.ts
@@ -1,5 +1,3 @@
-"use client";
-
 import {
   announceEip6963Provider,
   createEIP1193Provider,