+++ packages/eip6963/src/index.ts
@@ -1,3 +1,1 @@
-"use client";
-
 export { createSophonEIP6963Emitter } from "./emitter";